requirements.txt
streamlit
google-generativeai
SpeechRecognition
pillow
pyaudio
python-dotenv
